+++ crates/sevenmark_parser/src/parser/macro/macro_dday.rs
@@ -1,5 +1,5 @@
 use crate::parser::ParserInput;
-use sevenmark_ast::{AgeElement, Element, Span};
+use sevenmark_ast::{DdayElement, Element, Span};
 use winnow::Result;
 use winnow::combinator::delimited;
 use winnow::prelude::*;
@@ -8,14 +8,14 @@ use winnow::token::literal;
 
 use super::utils::parse_date;
 
-pub fn macro_age_parser(parser_input: &mut ParserInput) -> Result<Element> {
+pub fn macro_dday_parser(parser_input: &mut ParserInput) -> Result<Element> {
     let start = parser_input.current_token_start();
 
-    let date = delimited(literal("[age("), parse_date, literal(")]")).parse_next(parser_input)?;
+    let date = delimited(literal("[dday("), parse_date, literal(")]")).parse_next(parser_input)?;
 
     let end = parser_input.previous_token_end();
 
-    Ok(Element::Age(AgeElement {
+    Ok(Element::Dday(DdayElement {
         span: Span { start, end },
         date,
     }))